modularización funcionalidad button votes
